% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_setbacks.R
\name{add_setbacks}
\alias{add_setbacks}
\title{Add Seetbacks}
\usage{
add_setbacks(tidyparcel, tidydistrict, tidybuilding)
}
\arguments{
\item{tidyparcel}{A tidyparcel object is an simple features object depicting each side of a parcel and its label.}

\item{tidydistrict}{The tidydistrict corresponding to the tidyparcel. A tidydistrict object is one row from a tidyzoning simple features object.}

\item{tidybuilding}{A tidybuilding object is a special features object representing a building.}
}
\value{
Tidyparcel with a "setbacks" and "units" column
}
\description{
Takes in a tidyparcel, tidydistrict, and tidybuilding.
Returns the tidyparcel object with new columns to describing the setback and units of each side.
}
