% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_bedrooms.R
\name{check_bedrooms}
\alias{check_bedrooms}
\title{Compare building bedrooms per unit and allowed bedrooms}
\usage{
check_bedrooms(tidybuilding, tidydistrict, tidyparcel = NULL)
}
\arguments{
\item{tidybuilding}{A tidybuilding is a list of data frames used to represent a building.}

\item{tidydistrict}{The tidydistrict corresponding to the tidyparcel. A tidydistrict object is one row from a tidyzoning simple features object.}

\item{tidyparcel}{A tidyparcel object is an simple features object depicting each side of a parcel and its label (front, Interior side, Exterior side, rear, centroid).}
}
\value{
Returns TRUE or FALSE stating whether or not the building would be allowed in the district based on the bedrooms in each unit.
}
\description{
`check_bedrooms()` takes a tidybuilding and a tidydistrict to see if the district's zoning code allows the tidybuilding based on number of bedrooms.
}
