% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_fl_area.R
\name{check_fl_area}
\alias{check_fl_area}
\title{Compare building floor area and allowed floor area}
\usage{
check_fl_area(
  tidybuilding,
  tidydistrict = NULL,
  tidyparcel_dims = NULL,
  zoning_req = NULL
)
}
\arguments{
\item{tidybuilding}{A dataframe with fields to represent a building.
Created from the function \link{unify_tidybuilding}.}

\item{tidydistrict}{The tidydistrict corresponding to the tidyparcel.
A tidydistrict object is one row from a tidyzoning simple features object.}

\item{tidyparcel_dims}{A tidyparcel object is an simple features object
depicting each parcel and its its dimentions.}

\item{zoning_req}{The data frame result from the \code{get_zoning_req()} function.
If provided, the tidydistrict and parccel need not be provided.}
}
\value{
Returns TRUE, FALSE, or MAYBE stating whether or not the building would be allowed in the district based on floor area.
}
\description{
\code{check_fl_area()} takes a tidybuilding and a tidydistrict to see if the district's zoning code allows the tidybuilding based on floor area.
}
