% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_floors.R
\name{check_floors}
\alias{check_floors}
\title{Compare building floor count and allowed floors}
\usage{
check_floors(
  tidybuilding,
  tidydistrict = NULL,
  tidyparcel_dims = NULL,
  zoning_req = NULL
)
}
\arguments{
\item{tidybuilding}{A tidybuilding is a list of data frames used to represent a building.}

\item{tidydistrict}{The tidydistrict corresponding to the tidyparcel. A tidydistrict object is one row from a tidyzoning simple features object.}
}
\value{
Returns TRUE or FALSE stating whether or not the building would be allowed in the district based on total floors.
}
\description{
`check_floors()` takes a tidybuilding and a tidydistrict to see if the district's zoning code allows the tidybuilding based on number of floors.
}
