% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_footprint.R
\name{check_footprint}
\alias{check_footprint}
\title{Check if footprint fits in buildable area}
\usage{
check_footprint(tidybuilding, buildable_area, crs = 3081)
}
\arguments{
\item{tidybuilding}{A list of data frames with attributes representing a building.}

\item{buildable_area}{A geometry. Usually of the parcels buildable area calculated from the setback requirements.}

\item{crs}{The projected Coordinate Reference System for the study area. Must be in meters.}
}
\value{
Returns TRUE of FALSE stating whether or not the building footprint would fit in the buildable area.
}
\description{
`check_footprint()` takes a building footprint from tidybuilding object and checks to see if it will fit in the provided buildable area. The buildable area can be any shape.
}
