% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_footprint_area.R
\name{check_footprint_area}
\alias{check_footprint_area}
\title{Check Footprint Area}
\usage{
check_footprint_area(tidybuilding, tidyparcel_dims)
}
\arguments{
\item{tidybuilding}{A list of data frames describing a building}

\item{tidyparcel_dims}{A data frame with the dimmensions of each parcel}
}
\value{
The tidyparcel data frame with an added column stating whether the area of the building is smaller than the area of the parcel
}
\description{
Check Footprint Area
}
