% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_footprint_simple.R
\name{check_footprint_simple}
\alias{check_footprint_simple}
\title{Check if simple footprint fits in buildable area}
\usage{
check_footprint_simple(buildable_area, bldg_width_x, bldg_length_y)
}
\arguments{
\item{buildable_area}{A shape depicting the buildable area of a parcel. This is produced by the [get_buildable_area()] function.}

\item{bldg_width_x}{The width (in feet) of the building footprint}

\item{bldg_length_y}{The length (in feet) of the building footprint}
}
\value{
Returns TRUE of FALSE stating whether or not the building footprint would fit in the buildable area.
}
\description{
`check_footprint_simple()` takes a rectangular building footprint and checks to see if it will fit in the provided buildable area. The buildable area can be any shape.
}
