% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_height.R
\name{check_height}
\alias{check_height}
\title{Compare building height and allowed height}
\usage{
check_height(tidybuilding, tidydistrict)
}
\arguments{
\item{tidybuilding}{A tidybuilding is a list of data frames used to represent a building.}

\item{tidydistrict}{The tidydistrict corresponding to the tidyparcel. A tidydistrict object is one row from a tidyzoning simple features object.}
}
\value{
Returns TRUE or FALSE stating whether or not the building would be allowed in the district based on building height.
}
\description{
`check_height()` takes a tidybuilding and a tidydistrict to see if the district's zoning code allows the tidybuilding based on height.
}
