% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_land_use.R
\name{check_land_use}
\alias{check_land_use}
\title{Compare building land use and allowed land uses}
\usage{
check_land_use(tidybuilding, tidydistrict)
}
\arguments{
\item{tidybuilding}{A tidybuilding object is a special features object representing a building.}

\item{tidydistrict}{The tidydistrict corresponding to the tidyparcel. A tidydistrict object is one row from a tidyzoning simple features object.}
}
\value{
Returns TRUE or FALSE stating whether or not the building would be allowed in the district based on land use.
}
\description{
`check_land_use()` takes a tidybuilding and a tidydistrict to see if the district's zoning code allows the tidybuilding based on land use.
}
