% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_land_use_exceptions.R
\name{check_land_use_exceptions}
\alias{check_land_use_exceptions}
\title{Compare building land use and allowed land uses}
\usage{
check_land_use_exceptions(tidybuilding, tidydistrict)
}
\arguments{
\item{tidybuilding}{A tidybuilding is a list of data frames used to represent a building.}

\item{tidydistrict}{The tidydistrict corresponding to the tidyparcel. A tidydistrict object is one row from a tidyzoning simple features object.}
}
\value{
Returns TRUE or FALSE stating whether or not the building would be allowed in the district based on land use.
It will also return TRUE if the tidyzoning records a potential exception would allow the land use.
Note: If there is no recorded land use requirement in zoning code, it returns FALSE
}
\description{
`check_land_use_exceptions()`, like `check_land_use()`, will see if the district's zoning code allows the tidybuilding based on land use.
But this function also return TRUE if the land use is recorded in the tidyzoning as a potential exception.
}
