% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_lot_coverage.R
\name{check_lot_coverage}
\alias{check_lot_coverage}
\title{Compare building lot coverage to zoning requirements}
\usage{
check_lot_coverage(
  tidybuilding,
  tidydistrict = NULL,
  tidyparcel_dims,
  zoning_req = NULL
)
}
\arguments{
\item{zoning_req}{A data frame with the min_value and max_value of each
constraint. The result of the \link{get_zoning_req} function.}
}
\value{
Returns TRUE, FALSE, or MAYBE stating whether or not the building would be allowed in the district based on lot coverage.
}
\description{
\code{check_lot_coverage()} takes a tidybuilding, tidydistrict, and tidyparcel to see if the district's zoning code allows the tidybuilding based on lot coverage.
}
