% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_lot_size.R
\name{check_lot_size}
\alias{check_lot_size}
\title{Compare lot size to zonign requirements}
\usage{
check_lot_size(
  tidybuilding = NULL,
  tidydistrict = NULL,
  tidyparcel_dims,
  zoning_req = NULL
)
}
\arguments{
\item{tidybuilding}{A dataframe with fields to represent a building.
Created from the function \link{unify_tidybuilding}.}

\item{tidydistrict}{The tidydistrict corresponding to the tidyparcel.
A tidydistrict object is one row from a tidyzoning simple features object.}

\item{tidyparcel_dims}{A tidyparcel object is an simple features object
depicting each parcel and its its dimentions.}

\item{zoning_req}{The data frame result from the \code{get_zoning_req()} function.
If provided, the tidybuilding and tidydistrict need not be provided.}
}
\description{
Compare lot size to zonign requirements
}
