% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_parking_enclosed.R
\name{check_parking_enclosed}
\alias{check_parking_enclosed}
\title{Compare enclosed parking spaces to zoning requirements}
\usage{
check_parking_enclosed(
  tidybuilding,
  tidydistrict = NULL,
  tidyparcel_dims = NULL,
  zoning_req = NULL
)
}
\arguments{
\item{tidybuilding}{A dataframe with fields to represent a building.
Created from the function \link{unify_tidybuilding}.}

\item{tidydistrict}{The tidydistrict corresponding to the tidyparcel.
A tidydistrict object is one row from a tidyzoning simple features object.}

\item{tidyparcel_dims}{A tidyparcel object is an simple features object
depicting each parcel and its its dimentions.}

\item{zoning_req}{The data frame result from the \code{get_zoning_req()} function.
If provided, the tidydistrict and parccel need not be provided.}
}
\description{
Compare enclosed parking spaces to zoning requirements
}
