% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_unit_density.R
\name{check_unit_density}
\alias{check_unit_density}
\title{Compare building's unit density and allowed unit_density}
\usage{
check_unit_density(tidybuilding, tidydistrict)
}
\arguments{
\item{tidybuilding}{A tidybuilding is a list of data frames used to represent a building.}

\item{tidydistrict}{The tidydistrict corresponding to the tidyparcel. A tidydistrict object is one row from a tidyzoning simple features object.}
}
\value{
Returns TRUE or FALSE stating whether or not the building would be allowed in the district based on unit density.
}
\description{
`check_unit_density()` takes a tidybuilding and a tidydistrict to see if the district's zoning code allows the tidybuilding based on unit density.
}
