% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_unit_qty.R
\name{check_unit_qty}
\alias{check_unit_qty}
\title{Do total building units meet zoning requirements}
\usage{
check_unit_qty(
  tidybuilding,
  tidydistrict = NULL,
  tidyparcel_dims = NULL,
  zoning_req = NULL
)
}
\arguments{
\item{zoning_req}{}
}
\description{
`check_unit_qty` compares the total units in a tidybuilding object
 to the allowable number of units according to the tidyzoning data
 and returns TRUE, FALSE, or MAYBE
}
