% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_bldg_type.R
\name{find_bldg_type}
\alias{find_bldg_type}
\title{Find Building Type}
\usage{
find_bldg_type(tidybuilding)
}
\arguments{
\item{tidybuilding}{Tidybuilding data formatted in a data fram}
}
\value{
A string stating the building land use (1_family, 2_family, etc.).
This will coincide with the way it is written in the tidyzining object.
}
\description{
Takes in a tidybuilding and outputs the land use.
}
