% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_district_idx.R
\name{find_district_idx}
\alias{find_district_idx}
\title{Find District Index}
\usage{
find_district_idx(tidyparcel, tidyzoning)
}
\arguments{
\item{tidyparcel}{An sf data frame with a linestring for each side of the parcel.}

\item{tidyzoning}{The sf data frame with district geometry.}
}
\value{
The row number of the district to which the parcel belongs.
Returns NA if there is no district found.
}
\description{
Takes in an sf data frame with district geometry and a parcel.
Returns the row number of the district to which the parcel belongs.
}
