% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_buildable_area.R
\name{get_buildable_area}
\alias{get_buildable_area}
\title{Create a parcel's buildable area}
\usage{
get_buildable_area(tidyparcel_with_setbacks)
}
\arguments{
\item{tidyparcel_with_setbacks}{A tidyparcel object that has setback infomration added. tidyparcel_with_setbacks is the output of the [add_setbacks()] function.}
}
\value{
Returns a polygon representing the buildable area of the parcel.
}
\description{
get_buildable_area() takes a tidyparcel object with setback information and produces a polygon representing the builablel area of the parcel.
}
