% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_crs.R
\name{get_crs}
\alias{get_crs}
\title{Find EPSG code for State Plane CRS}
\usage{
get_crs(geom_data, large_area = FALSE)
}
\arguments{
\item{geom_data}{Either a simple feature collection or a
file path to a \code{.geojson} file.}

\item{large_area}{Set this to \code{TRUE} if your data may
cross multiple state planes. It will be a bit longer but
find the state plane that covers it the best}
}
\value{
The appropriate epsg code as an integer
}
\description{
\code{get_crs()} uses state plane geometry to find an appropriate NAD83
crs for the input geometry
}
\details{
The state plane data was compiled using \href{https://hub.arcgis.com/datasets/esri::usa-state-plane-zones-nad83/explore}{ArcGIS Hub}
and the \href{https://epsg.io/}{epsg.io} website.
}
