% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_raw_zoning_req.R
\name{get_raw_zoning_req}
\alias{get_raw_zoning_req}
\title{Title}
\usage{
get_raw_zoning_req(tidyzoning)
}
\arguments{
\item{tidyzoning}{A tidyzoning district}
}
\value{
a data frame with the zoning requirements in that districts
}
\description{
Title
}
