% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_tidyparcel_geo.R
\name{get_tidyparcel_geo}
\alias{get_tidyparcel_geo}
\title{Isolate known side geometry from tidyparcel data}
\usage{
get_tidyparcel_geo(parcels_data)
}
\arguments{
\item{parcels_data}{Either a file path to the a parcels geojson file or an sf object with the same data.}
}
\value{
an sf object with only the side geometries of the parcels with identified sides
}
\description{
Isolate known side geometry from tidyparcel data
}
