% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_unit_info.R
\name{get_unit_info}
\alias{get_unit_info}
\title{Create data frame with unit info}
\usage{
get_unit_info(file_path = NULL, string = NULL)
}
\value{
a data frame with information for each unit in the building
}
\description{
`get_unit_info()` takes a json representing a building and
returns a data frame with data for each unit in the building.
}
