% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_zoning_req.R
\name{get_zoning_req}
\alias{get_zoning_req}
\title{List district zoning requirement values}
\usage{
get_zoning_req(tidybuilding, tidydistrict, tidyparcel_dims)
}
\arguments{
\item{tidybuilding}{A tidybuilding is a list of data frames used to represent a building.}

\item{tidydistrict}{The tidydistrict corresponding to the tidyparcel. A tidydistrict object is one row from a tidyzoning simple features object.}

\item{tidyparcel_dims}{The simple features object with each parcel centroid and the parcel dimensions}
}
\value{
Returns a data frame with the value each zoning requirement for that specific building, parcel, and zoning district.
NA values indicate the requirement isn't recorded in that district.
If every value is NA, it could indicate that the building land use is not allowed in the zoning district.
}
\description{
Because many zoning requirements depend on the proposed building
or the parcel of that zoning district, \verb{the get_zoning_req()} function
takes a tidybuilding, a tidyparcel, and a tidydistrict and outputs a data
frame listing the set zoning requirements that those three objects would create.
If every value is NA, it could indicate that the building
land use is not allowed in the zoning district.
}
