% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ozfs_add_req.R
\name{ozfs_add_req}
\alias{ozfs_add_req}
\title{Add Zoning Requirements}
\usage{
ozfs_add_req(
  ozfs_file,
  district,
  constraint_type = "structure_constraints",
  constraint,
  units = NULL,
  land_uses,
  min_or_max,
  expression,
  overwrite = FALSE
)
}
\arguments{
\item{ozfs_file}{The file path for the ozfs file that needs editing}

\item{district}{The district that needs editing}

\item{constraint_type}{The constraint type. "structure_constraints", "lot_constraints", or "other_constraints"}

\item{constraint}{The constraint that you are either editing or adding.}

\item{units}{The units of the constraint. This is only necessary when you are adding a new constraint.}

\item{land_uses}{The land use that the constraint applies to. If it is an existing constraint, the land_uses has to match the existing land use possibilities.}

\item{min_or_max}{Whether the expression will be "min" or "max"}

\item{expression}{The expression used to describe the new zoning constraint.}

\item{overwrite}{If TRUE, will overwrite the existing file.}
}
\value{
A statement saying which file the updates were added to.
}
\description{
`ozfs_add_req()` takes an existing geojson file in ozfs format, and adds a simple constraint.
}
