% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ozfs_is_valid.R
\name{ozfs_is_valid}
\alias{ozfs_is_valid}
\title{Is it formatted to OZFS standards?}
\usage{
ozfs_is_valid(list_of_files)
}
\arguments{
\item{list_of_files}{A list of file paths to OZFS *.zoning files}
}
\value{
TRUE or FALSE stating whether data is in valid OZFS format or not
}
\description{
\code{ozfs_is_valid()} TRUE or FALSE stating whether or not the files are
properly formatted according to OZFS standards. Use \link{ozfs_validate} for
a detailed list of why files don't meet the standard.
}
