% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ozfs_validate.R
\name{ozfs_validate}
\alias{ozfs_validate}
\title{Find invalid ozfs zoning files}
\usage{
ozfs_validate(list_of_files)
}
\arguments{
\item{list_of_files}{A list of .zoning file paths}
}
\value{
A list of files names with ozfs errors or warnings
}
\description{
\code{ozfs_validate()} lists any errors in the formatting of the OZFS data
}
