% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show_shapes.R
\name{show_shapes}
\alias{show_shapes}
\title{Quickly visualize sf objects}
\usage{
show_shapes(shape1, shape2 = NA, shape3 = NA, shape4 = NA)
}
\arguments{
\item{shape1}{An sf object}

\item{shape2}{An sf object}

\item{shape3}{An sf object}

\item{shape4}{An sf object}
}
\value{
Returns a simple plot with the different shapes, each a different color
}
\description{
A quick way to visualize up to four different shapes using ggplot
}
