% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{tidybuilding_12fam}
\alias{tidybuilding_12fam}
\title{12-family tidybuilding}
\format{
An object of class \code{sf} (inherits from \code{data.frame}) with 1 rows and 8 columns.
}
\usage{
tidybuilding_12fam
}
\description{
A tidybuilding object depicting a 12-family home.
}
\keyword{datasets}
