% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{tidybuilding_4fam_thin}
\alias{tidybuilding_4fam_thin}
\title{A 4-family residence as a tidyzoning object}
\format{
A special features object with 1 row and 7 columns:
\describe{
  \item{units_2bed}{Number of 2-bedroom units in the building}
  \item{total_floors}{Floor count of the building}
  \item{max_unit_size}{Square footage of the smallest unit in the building}
  \item{min_unit_size}{Square footage of the biggest unit in the building}
  \item{floor_area}{Total floor area of the building}
  \item{building_height}{Building height}
  \item{geometry}{Building footprint geometry}

}
}
\usage{
tidybuilding_4fam_thin
}
\description{
This was created based on observations of a typical 4-family residence with a rectangular footprint.
}
\keyword{datasets}
