% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tz_all_checks.R
\name{tz_all_checks}
\alias{tz_all_checks}
\title{Run initial check functions}
\usage{
tz_all_checks(
  tidybuilding,
  tidyzoning,
  tidyparcel_dims,
  tidyparcel_geo = NULL,
  detailed_check = FALSE,
  run_check_land_use = TRUE,
  run_check_height = TRUE,
  run_check_height_eave = TRUE,
  run_check_floors = TRUE,
  run_check_unit_size = TRUE,
  run_check_far = TRUE,
  run_check_unit_density = TRUE,
  run_check_lot_coverage = TRUE,
  run_check_fl_area = TRUE,
  run_check_unit_qty = TRUE,
  run_check_footprint = FALSE
)
}
\arguments{
\item{tidybuilding}{A tidybuilding is a list of data frames used to represent a building.}

\item{tidyzoning}{A tidyzoning simple features object representing the zoning districts and regulations for an area.}

\item{tidyparcel_with_dimensions}{A data frame with each tidyparcel and its information. The zoning_id column in the tidyparcel_with_dimensions must correspond to the row numbers of the tidyzoning object.}

\item{tidyparcel_with_side_geom}{A data frame with each parcel side and its geometry}

\item{func_names}{A character list of the check functions you want to run.}

\item{run_parallel}{Logical. When `TRUE` the code runs on multiple cores (one less than your machine has available). Will save time if processing over 50,000 parcels. Default is `FALSE`}
}
\value{
A data frame with initial check function results.
}
\description{
`tz_initial_checks()` runs the initial tidyzoning check functions to see if certain attributes of the building are allowed in the parcel.
}
