% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unify_tidybuilding.R
\name{unify_tidybuilding}
\alias{unify_tidybuilding}
\title{Get summarized building info in one table}
\usage{
unify_tidybuilding(file_path = NULL, string = NULL)
}
\arguments{
\item{file_path}{a json file with ozfs building attributes see "insert hyperlink here"}

\item{string}{a json string with ozfs building attributes see "insert hyperlink here"}
}
\value{
a one-row data frame with values describing the building.
}
\description{
Get summarized building info in one table
}
