% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zoning_analysis_pipeline.R
\name{zoning_analysis_pipline}
\alias{zoning_analysis_pipline}
\title{Complete analysis to find where the building fits}
\usage{
zoning_analysis_pipline(
  bldg_file,
  parcels_file,
  ozfs_zoning_file,
  detailed_check = FALSE,
  run_check_land_use = TRUE,
  run_check_height = TRUE,
  run_check_height_eave = TRUE,
  run_check_floors = TRUE,
  run_check_unit_size = TRUE,
  run_check_far = TRUE,
  run_check_unit_density = TRUE,
  run_check_lot_coverage = TRUE,
  run_check_fl_area = TRUE,
  run_check_unit_qty = TRUE,
  run_check_lot_size = TRUE,
  run_check_parking_enclosed = TRUE,
  run_check_footprint = FALSE,
  crs_m = 3081
)
}
\arguments{
\item{bldg_file}{The path to the json file representing a building}

\item{parcels_file}{The path to the geojson file representing the parcels}

\item{ozfs_zoning_file}{The path to the geojson file with the ozfs zoning codes.}

\item{detailed_check}{When TRUE, every parcel passes through each check no matter the result,
and it take more time. When FALSE, subsequent checks are skipped as soon as one check reads FALSE}

\item{run_check_land_use}{Should the analysis run the check_land_use function? (logical)}

\item{run_check_height}{Should the analysis run the check_height function? (logical)}

\item{run_check_height_eave}{Should the analysis run the check_height_eave function? (logical)}

\item{run_check_floors}{Should the analysis run the check_floors function? (logical)}

\item{run_check_unit_size}{Should the analysis run the check_unit_size function? (logical)}

\item{run_check_far}{Should the analysis run the check_far function? (logical)}

\item{run_check_unit_density}{Should the analysis run the check_unit_density function? (logical)}

\item{run_check_lot_coverage}{Should the analysis run the check_lot_coverage function? (logical)}

\item{run_check_fl_area}{Should the analysis run the check_fl_area function? (logical)}

\item{run_check_unit_qty}{Should the analysis run the check_unit_qty function? (logical)}

\item{run_check_lot_size}{Should the analysis run the check_lot_size function? (logical)}

\item{run_check_parking_enclosed}{Should the analysis run the check_parking_enclosed function? (logical)}

\item{run_check_footprint}{Should the analysis run the check_footprint function? (logical)}

\item{crs_m}{Projected Coordinate Reference System (in meters) for your study area.}
}
\value{
a simple features data frame with the centroid of each parcel with a column
stating building allowance on the parcel and a column stating the reason
why certain parcels don't allow the building.
}
\description{
\code{zoning_analysis_pipline()} runs through all of the zoning checks to see which parcels
allow a certain building.
}
