% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zr_check_footprint_area.R
\name{zr_check_footprint_area}
\alias{zr_check_footprint_area}
\title{Is building footprint larger than parcel area?}
\usage{
zr_check_footprint_area(tidybuilding, tidyparcel_dims)
}
\arguments{
\item{tidybuilding}{A list of data frames describing a building}

\item{tidyparcel_dims}{A data frame with the dimmensions of each parcel}
}
\value{
The tidyparcel data frame with an added column stating whether the area of the building is smaller than the area of the parcel
}
\description{
Is building footprint larger than parcel area?
}
