% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zr_check_unit.R
\name{zr_check_unit}
\alias{zr_check_unit}
\title{Compare unit sizes and allowed unit sizes}
\usage{
zr_check_unit(
  tidybuilding,
  tidydistrict,
  tidyparcel_dims,
  building_json,
  zoning_req = NULL
)
}
\arguments{
\item{building_json}{The file path or json string of the json representing the building.}

\item{zoning_req}{The data frame result from the \code{get_zoning_req()} function.}
}
\value{
Returns TRUE, FALSE, or MAYBE stating whether or not the building would be allowed in the district based on the unit sizes.
}
\description{
\code{zr_check_unit()} takes a tidybuilding and a tidydistrict to see if the district's zoning code allows the tidybuilding based on unit sizes.
}
