% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_fl_area.R
\name{check_fl_area}
\alias{check_fl_area}
\title{Compare building floor area and allowed floor area}
\usage{
check_fl_area(tidybuilding, tidydistrict)
}
\arguments{
\item{tidybuilding}{A tidybuilding is a list of data frames used to represent a building.}

\item{tidydistrict}{The tidydistrict corresponding to the tidyparcel. A tidydistrict object is one row from a tidyzoning simple features object.}
}
\value{
Returns TRUE or FALSE stating whether or not the building would be allowed in the district based on floor area.
}
\description{
`check_fl_area()` takes a tidybuilding and a tidydistrict to see if the district's zoning code allows the tidybuilding based on floor area.
}
