% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_fl_area.R
\name{check_fl_area}
\alias{check_fl_area}
\title{Compare building floor area and allowed floor area}
\usage{
check_fl_area(vars, zoning_req)
}
\arguments{
\item{vars}{A data frame with a column for each OZFS variable. The result
of the \link{get_variables} function.}

\item{zoning_req}{A data frame with the min_value and max_value of each
constraint. The result of the \link{get_zoning_req} function.}
}
\value{
Returns "TRUE", "FALSE", or "MAYBE". "MAYBE" indicates a complex condition
the OZFS is currently not able to support.
}
\description{
\code{check_fl_area()} states whether a building is allowed on a parcel based
on its floor area. It returns TRUE, FALSE, or "MAYBE"
}
