% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_floors.R
\name{check_floors}
\alias{check_floors}
\title{Compare building floor count and allowed floors}
\usage{
check_floors(vars, zoning_req)
}
\arguments{
\item{vars}{A data frame with a column for each OZFS variable. The result
of the \link{get_variables} function.}

\item{zoning_req}{A data frame with the min_value and max_value of each
constraint. The result of the \link{get_zoning_req} function.}
}
\value{
Returns "TRUE", "FALSE", or "MAYBE". "MAYBE" indicates a complex condition
the OZFS is currently not able to support.
}
\description{
\code{check_floors()} states whether a building is allowed on a parcel based
on number of floors. It returns TRUE, FALSE, or "MAYBE"
}
