% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_floors.R
\name{check_floors}
\alias{check_floors}
\title{Compare building floor count and allowed floors}
\usage{
check_floors(
  tidybuilding,
  tidydistrict = NULL,
  tidyparcel_dims = NULL,
  zoning_req = NULL
)
}
\arguments{
\item{tidybuilding}{A dataframe with fields to represent a building.
Created from the function \link{unify_tidybuilding}.}

\item{tidydistrict}{The tidydistrict corresponding to the tidyparcel.
A tidydistrict object is one row from a tidyzoning simple features object.}

\item{tidyparcel_dims}{A tidyparcel object is an simple features object
depicting each parcel and its its dimentions.}

\item{zoning_req}{The data frame result from the \code{get_zoning_req()} function.
If provided, the tidydistrict and parccel need not be provided.}
}
\value{
Returns TRUE or FALSE stating whether or not the building would be allowed in the district based on total floors.
}
\description{
\code{check_floors()} takes a tidybuilding and a tidydistrict to see if the district's zoning code allows the tidybuilding based on number of floors.
}
