% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_footprint.R
\name{check_footprint}
\alias{check_footprint}
\title{Check if footprint fits in buildable area}
\usage{
check_footprint(tidybuilding, buildable_area, num_rotations = 24)
}
\arguments{
\item{tidybuilding}{A tidybuilding is a list of data frames used to represent a building.}

\item{buildable_area}{A shape depicting the buildable area of a parcel. This is produced by the [get_buildable_area()] function.}

\item{num_rotations}{The number of rotations you wish to spin the building footprint. More rotations will be more accurate, but will cost significantly more time.
Ex: if `num_rotations = 4` then the building footprint will be turned 90 degrees each time until it finds where it fits.
Ex: if `num_rotations = 24` then the building footprint will be turned 15 degrees each time until it finds where it fits.}
}
\value{
Returns TRUE of FALSE stating whether or not the building footprint would fit in the buildable area.
}
\description{
`check_footprint()` takes a building footprint from tidybuilding object and checks to see if it will fit in the provided buildable area. The buildable area can be any shape.
}
