% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_lot_coverage.R
\name{check_lot_coverage}
\alias{check_lot_coverage}
\title{Compare building lot coverage to zoning requirements}
\usage{
check_lot_coverage(tidybuilding, tidydistrict, tidyparcel)
}
\arguments{
\item{tidybuilding}{A tidybuilding object is a special features object representing a building.}

\item{tidydistrict}{The tidydistrict corresponding to the tidyparcel. A tidydistrict object is one row from a tidyzoning simple features object.}

\item{tidyparcel}{A tidyparcel object is an simple features object depicting each side of a parcel and its label (front, Interior side, Exterior side, rear).}
}
\value{
Returns TRUE or FALSE stating whether or not the building would be allowed in the district based on lot coverage.
}
\description{
`check_lot_coverage()` takes a tidybuilding, tidydistrict, and tidyparcel to see if the district's zoning code allows the tidybuilding based on lot coverage.
}
