% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_unit_density.R
\name{check_unit_density}
\alias{check_unit_density}
\title{Compare building's unit density and allowed unit_density}
\usage{
check_unit_density(
  tidybuilding,
  tidydistrict = NULL,
  tidyparcel_dims,
  zoning_req = NULL
)
}
\arguments{
\item{zoning_req}{A data frame with the min_value and max_value of each
constraint. The result of the \link{get_zoning_req} function.}
}
\value{
Returns TRUE or FALSE stating whether or not the building would be
allowed in the district based on unit density.
}
\description{
\code{check_unit_density()} takes a tidybuilding and a tidydistrict to see if
the district's zoning code allows the tidybuilding based on unit density.
}
