% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_unit_size.R
\name{check_unit_size}
\alias{check_unit_size}
\title{Compare unit sizes and allowed unit sizes}
\usage{
check_unit_size(tidybuilding, tidydistrict)
}
\arguments{
\item{tidybuilding}{A tidybuilding object is a special features object representing a building.}

\item{tidydistrict}{The tidydistrict corresponding to the tidyparcel. A tidydistrict object is one row from a tidyzoning simple features object.}
}
\value{
Returns TRUE or FALSE stating whether or not the building would be allowed in the district based on the unit sizes.
}
\description{
`check_unit_size()` takes a tidybuilding and a tidydistrict to see if the district's zoning code allows the tidybuilding based on unit sizes.
}
