% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_bldg_type.R
\name{find_bldg_type}
\alias{find_bldg_type}
\title{Find land use of a building}
\usage{
find_bldg_type(tidybuilding)
}
\arguments{
\item{tidybuilding}{A tidybuilding object is a special features object representing a building.}
}
\value{
A string stating the building land use (1_family, 2_family, etc.).
This will match the way land use is written in the tidyzining object.
}
\description{
`find_bldg_type()` takes in a tidybuilding and outputs the land use (1_family, 2_family, etc.).
}
