% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_district_idx.R
\name{find_district_idx}
\alias{find_district_idx}
\title{Find corresponding tidydistrict}
\usage{
find_district_idx(parcels_centroids_sf, tidyzoning, idx_col_name = "zoning_id")
}
\arguments{
\item{parcels_centroids_sf}{An sf object with the parcel centroids}

\item{tidyzoning}{A simple features object with a row for each zoning district and columns holding the geojson formatted zoning requirements of each district.}
}
\value{
The same parcels centroids object that was input but with an additional
column stating the row number of the corresponding tidyzoning district
}
\description{
`find_district_idx()` adds a zoning_id column to the parcels data that
contains the tidyzoning row index of the district the parcel is in
}
