% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_unit_info.R
\name{get_unit_info}
\alias{get_unit_info}
\title{Create data frame with unit info}
\usage{
get_unit_info(file_path = NULL, string = NULL)
}
\arguments{
\item{file_path}{a json file with ozfs building attributes}

\item{string}{a json string with ozfs building attributes}
}
\value{
a data frame with information for each unit in the building
}
\description{
\code{get_unit_info()} takes a json representing a building and
returns a data frame with data for each unit in the building.
}
