% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ozfs_is_valid.R
\name{ozfs_is_valid}
\alias{ozfs_is_valid}
\title{Check OZFS validity}
\usage{
ozfs_is_valid(list_of_files)
}
\arguments{
\item{...}{A file path to an OZFS geojson file or list of file paths, an sf object created from and OZFS geojson data, or a combination of those.}
}
\value{
TRUE or FALSE stating whether data is in valid OZFS format or not
}
\description{
\code{ozfs_is_valid()} State the validity of the ozfs data
}
