% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ozfs_validate.R
\name{ozfs_validate}
\alias{ozfs_validate}
\title{Validate ozfs}
\usage{
ozfs_validate(...)
}
\arguments{
\item{...}{A file path to an OZFS geojson file or list of file paths, an sf object created from and OZFS geojson data, or a combination of those.}
}
\value{
A data frame with info on where the errors were found.
}
\description{
`ozfs_validate()` lists any errors in the formatting of the OZFS data
}
