% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{tidybuilding_4fam_wide}
\alias{tidybuilding_4fam_wide}
\title{4-family tidybuilding}
\format{
An object of class \code{sf} (inherits from \code{data.frame}) with 1 rows and 7 columns.
}
\usage{
tidybuilding_4fam_wide
}
\description{
A tidybuilding object depicting a wider, 4-family home.
}
\keyword{datasets}
