% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tz_all_checks.R
\name{tz_all_checks}
\alias{tz_all_checks}
\title{Run initial check functions}
\usage{
tz_all_checks(
  tidybuilding,
  tidyzoning,
  tidyparcel_dims,
  tidyparcel_geo = NULL,
  detailed_check = FALSE,
  run_check_land_use = TRUE,
  run_check_height = TRUE,
  run_check_height_eave = TRUE,
  run_check_floors = TRUE,
  run_check_unit_size = TRUE,
  run_check_far = TRUE,
  run_check_unit_density = TRUE,
  run_check_lot_coverage = TRUE,
  run_check_fl_area = TRUE,
  run_check_unit_qty = TRUE,
  run_check_footprint = FALSE
)
}
\arguments{
\item{tidybuilding}{A tidybuilding is a list of data frames used to represent a building.}

\item{tidyzoning}{A tidyzoning simple features object representing the zoning districts and regulations for an area.}

\item{tidyparcel_dims}{A data frame with each tidyparcel and its information. The zoning_id column in the tidyparcel_with_dimensions must correspond to the row numbers of the tidyzoning object.}

\item{tidyparcel_geo}{A data frame with each parcel side and its geometry}

\item{detailed_check}{When TRUE, every parcel passes through each check no matter the result, and it take more time. When FALSE, subsequent checks are skipped as soon as one check reads FALSE}

\item{run_check_land_use}{Should the analysis run the check_land_use function}

\item{run_check_height}{Should the analysis run the check_height function}

\item{run_check_height_eave}{Should the analysis run the check_height_eave function}

\item{run_check_floors}{Should the analysis run the check_floors function}

\item{run_check_unit_size}{Should the analysis run the check_unit_size function}

\item{run_check_far}{Should the analysis run the check_far function}

\item{run_check_unit_density}{Should the analysis run the check_unit_density function}

\item{run_check_lot_coverage}{Should the analysis run the check_lot_coverage function}

\item{run_check_fl_area}{Should the analysis run the check_fl_area function}

\item{run_check_unit_qty}{Should the analysis run the check_unit_qty function}

\item{run_check_footprint}{Should the analysis run the check_footprint function}
}
\value{
A data frame with initial check function results.
}
\description{
\code{tz_initial_checks()} runs the initial tidyzoning check functions to see if certain attributes of the building are allowed in the parcel.
}
