% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unify_tidybuilding.R
\name{unify_tidybuilding}
\alias{unify_tidybuilding}
\title{Get summarized building info in one table}
\usage{
unify_tidybuilding(
  bldg_data_file = NULL,
  ozfs_data_file = NULL,
  bldg_data_string = NULL
)
}
\arguments{
\item{bldg_data_file}{a json file with ozfs building attributes see "insert hyperlink here"}

\item{ozfs_data_file}{the path to the}

\item{bldg_data_string}{a json bldg_data_string with ozfs building attributes see "insert hyperlink here"}
}
\value{
a one-row data frame with values describing the building.
}
\description{
Get summarized building info in one table
}
